+++ pages/customer/ToReceiveOrder.js
@@ -1,18 +1,17 @@
 import { Component } from "react";
 import { SafeAreaView, Text } from "react-native";
-import TabHeader from "../../components/TabHeader";
 
-class OrderCust extends Component {
+class ToReceiveOrder extends Component {
 
     render() {
 
         return (
 
             <SafeAreaView>
-                <Text>OrderCust</Text>
+                <Text>ToReceiveOrder</Text>
             </SafeAreaView>
         )
     }
 }
 
-export default OrderCust
+export default ToReceiveOrder